requirements.txt
python-dotenv
openai
requests
fastapi
uvicorn